Instance Update

+++ src/main/java/me/t3sl4/updater/utils/GeneralUtil.java
@@ -107,11 +107,13 @@ public class GeneralUtil {
     }
 
     public static void systemShutdown() {
-        try {
-            Path lockFilePath = Path.of(System.getProperty("user.home"), ".onder_grup_updater.pid");
-            Files.deleteIfExists(lockFilePath);
-        } catch (IOException e) {
-            e.printStackTrace();
+        if(!System.getProperty("os.name").toLowerCase().contains("win")) {
+            try {
+                Path lockFilePath = Path.of(System.getProperty("user.home"), ".onder_grup_updater.pid");
+                Files.deleteIfExists(lockFilePath);
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
         }
 
         Platform.exit();
